Upgraded OpenSSL to 1.1.1g. Made building simulator binaries controlled by an environment variable. Fixed builds on Mac OS X 10.15.7

+++ ios/bVNC/build-libs.sh
@@ -32,7 +32,7 @@ rsync -avP /opt/libjpeg-turbo/lib /opt/libjpeg-turbo/include ./bVNC.xcodeproj/li
 if git clone https://github.com/x2on/OpenSSL-for-iPhone.git
 then
   pushd OpenSSL-for-iPhone
-  ./build-libssl.sh --deprecated --version=1.1.1e
+  ./build-libssl.sh --deprecated --version=1.1.1g
   popd
 else
   echo "Found OpenSSL-for-iPhone directory, assuming it is built, please remove with 'rm -rf OpenSSL-for-iPhone' to rebuild"
@@ -64,26 +64,32 @@ then
   rm -rf build_simulator64
 fi
 
-if [ -n "${CLEAN}" -o ! -d build_simulator64 ]
+if [ -n "${SIMULATOR_BUILD}" -a ! -d build_simulator64 ]
 then
-  mkdir -p build_simulator64
-  pushd build_simulator64
-  cmake .. -G Xcode -DENABLE_BITCODE=OFF -DARCHS='arm64 x86_64' \
-    -DCMAKE_C_FLAGS='-D OPENSSL_MIN_API=0x00908000L -D OPENSSL_API_COMPAT=0x00908000L' \
-    -DOPENSSL_SSL_LIBRARY=../../OpenSSL-for-iPhone/lib/libssl.a \
-    -DOPENSSL_CRYPTO_LIBRARY=../../OpenSSL-for-iPhone/lib/libcrypto.a \
-    -DOPENSSL_INCLUDE_DIR=../../OpenSSL-for-iPhone/include \
-    -DCMAKE_INSTALL_PREFIX=./libs -DCMAKE_INSTALL_PREFIX=./libs \
-    -DCMAKE_TOOLCHAIN_FILE=../../ios-cmake/ios.toolchain.cmake \
-    -DPLATFORM=SIMULATOR64 -DCMAKE_XCODE_ATTRIBUTE_DEVELOPMENT_TEAM="${DEVELOPMENT_TEAM}" \
-    -DJPEG_LIBRARY=/opt/libjpeg-turbo/lib/libjpeg.a -DJPEG_INCLUDE_DIR=/opt/libjpeg-turbo/include \
-    -DBUILD_SHARED_LIBS=OFF -DENABLE_VISIBILITY=ON -DENABLE_ARC=OFF
-else
-  pushd build_simulator64
-fi
-cmake --build . --config ${TYPE} --target install
+  echo "Simulator build"
 
-popd
+  if [ ! -d build_simulator64 ]
+  then
+    mkdir -p build_simulator64
+    pushd build_simulator64
+    cmake .. -G Xcode -DENABLE_BITCODE=OFF -DARCHS='x86_64' \
+      -DCMAKE_C_FLAGS='-D OPENSSL_MIN_API=0x00908000L -D OPENSSL_API_COMPAT=0x00908000L' \
+      -DOPENSSL_SSL_LIBRARY=../../OpenSSL-for-iPhone/lib/libssl.a \
+      -DOPENSSL_CRYPTO_LIBRARY=../../OpenSSL-for-iPhone/lib/libcrypto.a \
+      -DOPENSSL_INCLUDE_DIR=../../OpenSSL-for-iPhone/include \
+      -DCMAKE_INSTALL_PREFIX=./libs -DCMAKE_INSTALL_PREFIX=./libs \
+      -DCMAKE_TOOLCHAIN_FILE=../../ios-cmake/ios.toolchain.cmake \
+      -DPLATFORM=SIMULATOR64 -DCMAKE_XCODE_ATTRIBUTE_DEVELOPMENT_TEAM="${DEVELOPMENT_TEAM}" \
+      -DJPEG_LIBRARY=/opt/libjpeg-turbo/lib/libjpeg.a -DJPEG_INCLUDE_DIR=/opt/libjpeg-turbo/include \
+      -DBUILD_SHARED_LIBS=OFF -DENABLE_VISIBILITY=ON -DENABLE_ARC=OFF \
+      -DDEPLOYMENT_TARGET=12.0 \
+      -DWITH_GCRYPT=OFF
+     popd
+  fi
+  pushd build_simulator64
+  cmake --build . --config ${TYPE} --target install
+  popd
+fi
 
 if [ -n "${CLEAN}" ]
 then
@@ -91,48 +97,60 @@ then
 fi
 
 echo 'PRODUCT_BUNDLE_IDENTIFIER = com.iiordanov.bVNC' > ${TYPE}.xcconfig
-if [ -n "${CLEAN}" -o ! -d build_iphone ]
+if [ -z "${SIMULATOR_BUILD}" ]
 then
-  mkdir -p build_iphone
+  echo "Non-simulator build"
+
+  if [ ! -d build_iphone ]
+  then
+    mkdir -p build_iphone
+    pushd build_iphone
+    cmake .. -G Xcode -DENABLE_BITCODE=OFF -DARCHS='arm64' \
+      -DCMAKE_C_FLAGS='-D OPENSSL_MIN_API=0x00908000L -D OPENSSL_API_COMPAT=0x00908000L' \
+      -DOPENSSL_SSL_LIBRARY=../../OpenSSL-for-iPhone/lib/libssl.a \
+      -DOPENSSL_CRYPTO_LIBRARY=../../OpenSSL-for-iPhone/lib/libcrypto.a \
+      -DOPENSSL_INCLUDE_DIR=../../OpenSSL-for-iPhone/include \
+      -DCMAKE_INSTALL_PREFIX=./libs -DCMAKE_INSTALL_PREFIX=./libs \
+      -DCMAKE_TOOLCHAIN_FILE=../../ios-cmake/ios.toolchain.cmake \
+      -DPLATFORM=OS64 -DCMAKE_XCODE_ATTRIBUTE_DEVELOPMENT_TEAM="${DEVELOPMENT_TEAM}" \
+      -DJPEG_LIBRARY=/opt/libjpeg-turbo/lib/libjpeg.a -DJPEG_INCLUDE_DIR=/opt/libjpeg-turbo/include \
+      -DBUILD_SHARED_LIBS=OFF -DENABLE_VISIBILITY=ON -DENABLE_ARC=OFF \
+      -DDEPLOYMENT_TARGET=12.0 \
+      -DWITH_GCRYPT=OFF
+    popd
+  fi
   pushd build_iphone
-  cmake .. -G Xcode -DENABLE_BITCODE=OFF -DARCHS='arm64 x86_64' \
-    -DCMAKE_C_FLAGS='-D OPENSSL_MIN_API=0x00908000L -D OPENSSL_API_COMPAT=0x00908000L' \
-    -DOPENSSL_SSL_LIBRARY=../../OpenSSL-for-iPhone/lib/libssl.a \
-    -DOPENSSL_CRYPTO_LIBRARY=../../OpenSSL-for-iPhone/lib/libcrypto.a \
-    -DOPENSSL_INCLUDE_DIR=../../OpenSSL-for-iPhone/include \
-    -DCMAKE_INSTALL_PREFIX=./libs -DCMAKE_INSTALL_PREFIX=./libs \
-    -DCMAKE_TOOLCHAIN_FILE=../../ios-cmake/ios.toolchain.cmake \
-    -DPLATFORM=OS -DCMAKE_XCODE_ATTRIBUTE_DEVELOPMENT_TEAM="${DEVELOPMENT_TEAM}" \
-    -DJPEG_LIBRARY=/opt/libjpeg-turbo/lib/libjpeg.a -DJPEG_INCLUDE_DIR=/opt/libjpeg-turbo/include \
-    -DBUILD_SHARED_LIBS=OFF -DENABLE_VISIBILITY=ON -DENABLE_ARC=OFF
-else
-  pushd build_iphone
-fi
 
-# Workaround for missing PRODUCT_BUNDLE_IDENTIFIER in generated LibVNCServer.xcodeproj/project.pbxproj file
-sed -i.bak '/ *ARCHS =.*/a\
-PRODUCT_BUNDLE_IDENTIFIER="com.iiordanov.bVNC";
-' LibVNCServer.xcodeproj/project.pbxproj
+  # Workaround for missing PRODUCT_BUNDLE_IDENTIFIER in generated LibVNCServer.xcodeproj/project.pbxproj file
+  sed -i.bak '/ *ARCHS =.*/a\
+  PRODUCT_BUNDLE_IDENTIFIER="com.iiordanov.bVNC";
+  ' LibVNCServer.xcodeproj/project.pbxproj
 
-cmake --build . --config ${TYPE} --target install
-
-# Workaround for a file that blocks the creation of a directory
-if [ -f ${TYPE}-iphoneos ]
-then
-  rm ${TYPE}-iphoneos
-  mkdir ${TYPE}-iphoneos
   cmake --build . --config ${TYPE} --target install
-fi
 
-popd
+  # Workaround for a file that blocks the creation of a directory
+  if [ -f ${TYPE}-iphoneos ]
+  then
+    rm ${TYPE}-iphoneos
+    mkdir ${TYPE}-iphoneos
+    cmake --build . --config ${TYPE} --target install
+  fi
+
+  popd
+fi
 
 rsync -avP build_iphone/libs/ libs_combined/
 pushd libs_combined
 
 for lib in lib/lib*.a
 do
- echo $lib
- lipo ../build_iphone/libs/${lib} ../build_simulator64/libs/${lib} -output ${lib} -create
+  echo $lib
+  if [ -z "${SIMULATOR_BUILD}" ]
+  then
+    lipo ../build_iphone/libs/${lib} -output ${lib} -create
+  else
+    lipo ../build_simulator64/libs/${lib} -output ${lib} -create
+  fi
 done
 
 popd
